refactor(platform): split feature-switch reader and writer modules
The feature-switch reader (`featureSwitch$`, `apiBackendEnabled$`, etc.)
must be importable from `fetch.ts` to derive the api base url. Routing
the writer (reload/set/reset commands) through `zeroClient$` therefore
forms a static cycle:

  feature-switch.ts -> api-client.ts -> fetch.ts -> feature-switch.ts

Move the writer commands to a new `update-feature-switch.ts` so the
reader file no longer imports the api client. Nothing imports the writer
back through the cycle, breaking it.
