+++ turbo/apps/platform/src/signals/external/update-feature-switch.ts
@@ -1,0 +1,79 @@
+import { command } from "ccstate";
+import { getAllFeatureStates } from "@vm0/core/feature-switch";
+import { zeroFeatureSwitchesContract } from "@vm0/api-contracts/contracts/zero-feature-switches";
+import { FeatureSwitchKey } from "@vm0/connectors/feature-switch-key";
+import { clerk$ } from "../auth";
+import { accept } from "../../lib/accept.ts";
+import { zeroClient$ } from "../api-client.ts";
+import { setFeatureSwitchLocalStorage$ } from "./feature-switch.ts";
+
+function applySwitches(
+  result: Record<FeatureSwitchKey, boolean>,
+  overrides: Partial<Record<string, boolean>> | undefined,
+) {
+  if (!overrides) {
+    return;
+  }
+  for (const key of Object.values(FeatureSwitchKey)) {
+    const value = overrides[key];
+    if (value !== undefined) {
+      result[key] = Boolean(value);
+    }
+  }
+}
+
+export const reloadFeatureSwitch$ = command(
+  async ({ get, set }, signal: AbortSignal) => {
+    const clerk = await get(clerk$);
+    signal.throwIfAborted();
+
+    const user = clerk.user;
+    if (!user) {
+      return;
+    }
+
+    const client = get(zeroClient$)(zeroFeatureSwitchesContract);
+    const result = await accept(
+      client.get({ fetchOptions: { signal } }),
+      [200],
+    );
+    signal.throwIfAborted();
+
+    const combined = getAllFeatureStates({
+      userId: user.id,
+      email: user.primaryEmailAddress?.emailAddress,
+      orgId: clerk.organization?.id,
+    });
+    applySwitches(combined, result.body.switches);
+
+    set(setFeatureSwitchLocalStorage$, JSON.stringify(combined));
+  },
+);
+
+export const setFeatureSwitch$ = command(
+  async (
+    { get, set },
+    overrides: Partial<Record<FeatureSwitchKey, boolean>>,
+    signal: AbortSignal,
+  ) => {
+    const client = get(zeroClient$)(zeroFeatureSwitchesContract);
+    await accept(
+      client.update({
+        body: { switches: overrides },
+        fetchOptions: { signal },
+      }),
+      [200],
+    );
+    signal.throwIfAborted();
+    await set(reloadFeatureSwitch$, signal);
+  },
+);
+
+export const resetFeatureSwitches$ = command(
+  async ({ get, set }, signal: AbortSignal) => {
+    const client = get(zeroClient$)(zeroFeatureSwitchesContract);
+    await accept(client.delete({ fetchOptions: { signal } }), [200]);
+    signal.throwIfAborted();
+    await set(reloadFeatureSwitch$, signal);
+  },
+);
